feat: move loading transactions to `Calendar`
